Neue testImplementation für das Testen der Controller mit Keycloak hinzugefügt und den ersten Test so bearbeitet, dass dieser den Keycloak Account mockt.
